pyproject.toml
[project]
name = "eyf"
version = "0.1.0"
description = ""
authors = [
    {name = "Franco Jofré",email = "francoj.jofre@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "pandas (>=2.3.2)",
    "numpy (>=2.2.0)",
    "matplotlib (>=3.0.0)",
    "scikit-learn (>=1.7.2)",
    "scipy (>=1.4.0)",
    "ipykernel (>=6.30.1)",
    "tqdm (>=4.67.1)",
    "seaborn (>=0.13.2)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "optuna (>=4.5.0,<5.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "PyYAML (>=6.0.0)"
]

[tool.poetry]
packages = [{include = "eyf", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
